Removed faulty LDR Fallback
Even though the maps of L4D2 are compiled in HDR mode only and HDR itself is never disabled, Valve still had in place fallbacks to the LDR versions of the skyboxes of the original 5 campaigns of L4D2 + The Passing for when people are playing in Low shader settings.
Problem is, the actual materials and textures of the LDR skyboxes aren't present in the game files resulting in flat colors for the sky at best and straight up missing textures at worst.